move subscripts tests in own file
